`Refactor: Restructure "drawdata" package and improve association update logic`
